qbank/qbank-backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
redis==5.0.8
rq==1.16.2
kafka-python==2.0.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.32
PyJWT==2.9.0
black==24.8.0
flake8==7.1.1
pytest==8.3.2
httpx==0.27.2